Launching & Positioning Haldiram’s in Switzerland
This presentation outlines the strategic plan for launching and positioning Haldiram’s in Switzerland. Our collaborative effort with the School of Management Fribourg (HEG-Fr) and insights from a Swiss team led to this comprehensive plan. We explored Swiss consumer preferences, identified market potential, and crafted a strategy that emphasizes Haldiram’s authenticity, quality, and appeal to health-conscious consumers. The goal is to introduce Swiss consumers to the rich flavors of Indian cuisine while maintaining Haldiram’s brand integrity.